+++ bin/install.js
@@ -460,114 +460,24 @@ function resolveSharedHooksDirName(runtime) {
   return name;
 }
 
-// #3897 rung 3 — the 16 roles HALT.md measured as widening under derivation
-// (declare Write/Edit, never in the pre-#3897 CODEX_AGENT_SANDBOX map, so the
-// old `|| 'read-only'` fallback silently under-granted them). Pinned to
-// `read-only` pending the open question of whether Codex enforces
-// `sandbox_mode` or treats it as advisory (HALT.md). This list is CLOSED and
-// SHRINK-ONLY: a new writing role never lands here (S6, T26); it is validated
-// against the live tool contract every time it is consulted
-// (`_deriveCodexSandboxMode` below) and against the real `agents/` directory
-// (`validateCodexSandboxHolds` below) so a stale or orphaned entry fails
-// loudly instead of being silently honored forever — the exact defect this
-// rung deletes, rebuilt one list later (HALT.md's ledger claim).
-const CODEX_SANDBOX_HOLDS = {
-  'gsd-ai-researcher': 'declares Write/Edit; pending Codex sandbox_mode enforcement decision',
-  'gsd-code-fixer': 'declares Write/Edit; pending Codex sandbox_mode enforcement decision',
-  'gsd-code-reviewer': 'declares Write/Edit; pending Codex sandbox_mode enforcement decision',
-  'gsd-debug-session-manager': 'declares Write/Edit; pending Codex sandbox_mode enforcement decision',
-  'gsd-doc-classifier': 'declares Write/Edit; pending Codex sandbox_mode enforcement decision',
-  'gsd-doc-synthesizer': 'declares Write/Edit; pending Codex sandbox_mode enforcement decision',
-  'gsd-doc-verifier': 'declares Write/Edit; pending Codex sandbox_mode enforcement decision',
-  'gsd-doc-writer': 'declares Write/Edit; pending Codex sandbox_mode enforcement decision',
-  'gsd-dom-verifier': 'declares Write/Edit; pending Codex sandbox_mode enforcement decision',
-  'gsd-domain-researcher': 'declares Write/Edit; pending Codex sandbox_mode enforcement decision',
-  'gsd-eval-auditor': 'declares Write/Edit; pending Codex sandbox_mode enforcement decision',
-  'gsd-eval-planner': 'declares Write/Edit; pending Codex sandbox_mode enforcement decision',
-  'gsd-intel-updater': 'declares Write/Edit; pending Codex sandbox_mode enforcement decision',
-  'gsd-pattern-mapper': 'declares Write/Edit; pending Codex sandbox_mode enforcement decision',
-  'gsd-ui-auditor': 'declares Write/Edit; pending Codex sandbox_mode enforcement decision',
-  'gsd-ui-researcher': 'declares Write/Edit; pending Codex sandbox_mode enforcement decision',
-};
-
-// #3897 rung 3 — true iff a `tools:` frontmatter value declares Write or Edit
-// as a whole token (never a substring match, so e.g. a hypothetical
-// "Edith"-named tool could never collide). Single predicate owner for both
-// the emitter (`_deriveCodexSandboxMode`) and the posture check
-// (`agent-install-check.cts`'s `checkCodexSandboxPosture`, which requires
-// this module lazily rather than re-implementing the rule).
-function _codexToolsDeclareWriteOrEdit(toolsRaw) {
-  const tokens = String(toolsRaw || '').split(',').map((t) => t.trim());
-  return tokens.includes('Write') || tokens.includes('Edit');
-}
-
-// #3897 rung 3 — the single owner of sandbox_mode derivation: `workspace-write`
-// iff the role's own frontmatter `tools:` declares Write/Edit, UNLESS the role
-// is held (CODEX_SANDBOX_HOLDS) at `read-only`. A held role whose live tool
-// contract no longer derives broader than its pin is a STALE hold (S4) —
-// fail loudly naming the role rather than silently honoring it forever,
-// which is exactly the hand-maintained-subset-map defect this rung deletes.
-//
-// #3897 security review follow-up: `agentName` here MUST be an identity the
-// content's own frontmatter cannot change — i.e. the source `.md` FILENAME
-// stem (what `validateCodexSandboxHolds` already verifies exists), never a
-// frontmatter-derived display name. The old scheme keyed this off frontmatter
-// `name:`, which let an edited (or merely recased) `name:` field silently
-// escape a hold: `CODEX_SANDBOX_HOLDS` is a SUBTRACTION from a derivation
-// that defaults to `workspace-write`, so a missed lookup fails OPEN, not
-// closed. The lookup is case-normalized here as a second line of defense
-// against a differently-cased identity reaching this function; callers are
-// still responsible for passing the real filename stem, not the frontmatter
-// field.
-function _deriveCodexSandboxModeFromFrontmatter(agentName, frontmatterText) {
-  const toolsRaw = extractFrontmatterField(frontmatterText, 'tools') || '';
-  const derivesBroader = _codexToolsDeclareWriteOrEdit(toolsRaw);
-  const holdKey = String(agentName == null ? '' : agentName).toLowerCase();
-  if (Object.prototype.hasOwnProperty.call(CODEX_SANDBOX_HOLDS, holdKey)) {
-    if (!derivesBroader) {
-      throw new Error(
-        `CODEX_SANDBOX_HOLDS: stale hold for "${agentName}" — its current tools: frontmatter no longer ` +
-        'declares Write/Edit, so the pin no longer holds anything broader than what derivation would ' +
-        'already produce. Remove this entry (ADR-3473 §8.3 / HALT.md: the hold list is self-invalidating ' +
-        'and must shrink to zero, never be silently honored once stale).'
-      );
-    }
-    return 'read-only';
-  }
-  return derivesBroader ? 'workspace-write' : 'read-only';
-}
-
-// Exported form for external callers (e.g. `checkCodexSandboxPosture`) that
-// hold raw agent markdown rather than an already-extracted frontmatter slice.
-// Parses via the SAME extractFrontmatterAndBody/extractFrontmatterField pair
-// generateCodexAgentToml uses — never a second, independent parser.
-//
-// `agentName` is a HOLD-LOOKUP IDENTITY, not a display name: callers must
-// pass the agent's canonical source filename stem (e.g. `gsd-doc-writer` for
-// `agents/gsd-doc-writer.md`), never a value read from the content's own
-// frontmatter `name:` field (see the security note on
-// `_deriveCodexSandboxModeFromFrontmatter` above).
-function deriveCodexSandboxMode(agentName, agentContent) {
-  const { frontmatter } = extractFrontmatterAndBody(agentContent);
-  return _deriveCodexSandboxModeFromFrontmatter(agentName, frontmatter || '');
-}
-
-// #3897 rung 3 (S5) — every CODEX_SANDBOX_HOLDS key must still name a real
-// `<agentsSrcDir>/<role>.md`. A hold for a role that no longer exists is
-// stale and must fail loudly, not be silently ignored (else the hold list
-// only ever grows/rots instead of shrinking to zero as HALT.md's decision is
-// taken). Called once per Codex install, before any TOML is written.
-function validateCodexSandboxHolds(agentsSrcDir) {
-  for (const role of Object.keys(CODEX_SANDBOX_HOLDS)) {
-    const agentFile = path.join(agentsSrcDir, `${role}.md`);
-    if (!fs.existsSync(agentFile)) {
-      throw new Error(
-        `CODEX_SANDBOX_HOLDS: stale hold for "${role}" — no ${agentFile} exists. The hold list is ` +
-        'closed and shrink-only (ADR-3473 §8.3 / HALT.md); remove this entry.'
-      );
-    }
-  }
-}
+// #3897 rung 3 — sandbox_mode derivation, the hold list, and the hold-roster
+// validator now live in `src/codex-agent-toml.cts` (compiled to
+// `gsd-core/bin/lib/codex-agent-toml.cjs`), NOT here. This module used to be
+// the sole owner, and `agent-install-check.cts`'s `checkCodexSandboxPosture`
+// lazily `require()`d THIS FILE to reach `deriveCodexSandboxMode` — but
+// requiring `bin/install.js` runs its whole top-level script, including the
+// CLI's ASCII banner print to stdout, which corrupted every stdout-JSON
+// caller downstream of that posture check (`gsd-tools validate agents`).
+// `codex-agent-toml.cjs` is a genuine leaf (no top-level side effects), so
+// both this file and `agent-install-check.cts` import the derivation from
+// there — ONE owner, no second predicate. See that module's header for the
+// full rationale, and CAUSE B (below, `installCodexConfig`) for the removal
+// of `validateCodexSandboxHolds`'s call from the install runtime path.
+const {
+  CODEX_SANDBOX_HOLDS,
+  deriveCodexSandboxMode,
+  validateCodexSandboxHolds,
+} = require(path.join(__dirname, '..', 'gsd-core', 'bin', 'lib', 'codex-agent-toml.cjs'));
 
 // Copilot tool name mapping — Claude Code tools to GitHub Copilot tools
 // Tool mapping applies ONLY to agents, NOT to skills (per CONTEXT.md decision)
@@ -4111,7 +4021,12 @@ function generateCodexAgentToml(agentName, agentContent, modelOverrides = null, 
   // 2). The former hand-maintained CODEX_AGENT_SANDBOX map is deleted (ADR-3473
   // §8.3): it was fully redundant with this derivation, zero disagreements
   // across all 11 entries. Never a silent `|| 'read-only'` fallback either.
-  const sandboxMode = _deriveCodexSandboxModeFromFrontmatter(agentName, frontmatterText);
+  // Derivation itself lives in codex-agent-toml.cjs, which does no frontmatter
+  // parsing of its own (no third copy of that extraction) — it takes the
+  // already-resolved `tools:` value, reused from this function's own
+  // `frontmatterText` extraction above rather than re-parsed a second time.
+  const toolsRaw = extractFrontmatterField(frontmatterText, 'tools') || '';
+  const sandboxMode = deriveCodexSandboxMode(agentName, toolsRaw);
   const resolvedName = extractFrontmatterField(frontmatterText, 'name') || agentName;
   const resolvedDescription = toSingleLine(
     extractFrontmatterField(frontmatterText, 'description') || `GSD agent ${resolvedName}`
@@ -7024,9 +6939,19 @@ function installCodexConfig(targetDir, agentsSrc, sandboxTier = 'codex-agent-san
   }
   fs.mkdirSync(agentsTomlDir, { recursive: true });
 
-  // #3897 rung 3 (S5) — fail loudly, before writing anything, if the hold
-  // list names a role that no longer exists in the bundled agents source.
-  validateCodexSandboxHolds(agentsSrc);
+  // #3897 rung 3 (CAUSE B fix) — validateCodexSandboxHolds is deliberately
+  // NOT called here. The "no stale holds" invariant is a REPO invariant about
+  // the canonical roster in `agents/`, not a property of whatever directory
+  // an install happens to read from: a partial or synthetic `agentsSrc` (a
+  // test fixture, a `--config-dir` subset) legitimately contains only a few
+  // agents, and a held role simply absent from THIS source dir must be
+  // inert, not fatal. Throwing here also masked unrelated failures further
+  // down this same loop (e.g. the name-injection/path-escape guard on
+  // `agentTomlPath` below), since this check ran first and unconditionally.
+  // The invariant is still enforced — as a test over the real `agents/`
+  // roster (tests/codex-config.test.cjs T24/T25) — just never on this
+  // runtime path. See src/codex-agent-toml.cts's `validateCodexSandboxHolds`
+  // docblock for the full rationale.
 
   const agentEntries = fs.readdirSync(agentsSrc).filter(f => f.startsWith('gsd-') && f.endsWith('.md'));
   const agents = [];
